Rotating the wheel
The wheel now rotates for 8 seconds, it does this by setting the class
to a rotating one temporarily, I may change the src to the spinning
wheel. The rotation doesn't work correctly at altered screen sizes when
the aspect ratio of the image is ruined. This will have to be resolved,
next task is the ball animation.