Current version
Tested on firefox, safari and chrome. Unfortunately unable to make
responsive, after attempting to get this working on mobile, I believe a
complete mobile version would be better. The ball drop animation had
some issues with changing the positions as the transform origin had to
be changed as well, separating the ball into it's own small image would
be ideal.