Fixed a few issues
Fixed a few bugs. A value needs to be entered prior to spinning the
wheel. The spin button can no longer be spammed. The character limit has
been set to 16 for the answers. The ball gradually rotates inwards to be
inside the correct column as well. The aesthetics of the pop up have
been improved as well.